feat: init plic

+++ src/kernel/memlayout.rs
@@ -21,11 +21,11 @@ use crate::riscv::*;
 
 // qemu puts UART registers here in physical memory.
 pub const UART0: usize = 0x10000000;
-pub const UART0_IRQ: u64 = 10;
+pub const UART0_IRQ: usize = 10;
 
 // virtio mmio interface
 pub const VIRTIO0: usize =  0x10001000;
-pub const VIRTIO0_IRQ: u64 =  1;
+pub const VIRTIO0_IRQ: usize =  1;
 
 // core local interruptor (CLINT), which contains the timer.
 pub const CLINT: u64 =  0x2000000;
